Fix duplicate newsletter article slugs.
This change allows newsletter articles to have a unique slug within a
newsletter issue. It is required to put %post_id% in the newsletter
article URL, otherwise the wrong article may be returned. In the future
rewrite logic may be added to recognize by post name and issue number,
so that the post id doesn't need to be in the URL anymore.